refactor(dispatch): split request handling flow (#0000)
