+++ crates/perl-lsp-rs/src/runtime/dispatch/mod.rs
@@ -18,7 +18,7 @@
 //!
 //! 1. Request arrives via JSON-RPC transport
 //! 2. Cancellation token registered for long-running operations
-//! 3. Method string matched to handler in `handle_request`
+//! 3. Method string matched to handler in `routing::route_request`
 //! 4. Handler invoked with params and optional request ID
 //! 5. Response returned (or None for notifications)
 //! 6. Cancellation token cleaned up
@@ -48,265 +48,30 @@ mod ast_explorer;
 mod cancellation;
 mod experimental;
 mod lifecycle;
+mod preflight;
 mod request_cancellation;
+mod response;
+mod routing;
 mod text_document;
 mod workspace;
 
-pub(crate) use cancellation::early_cancel_or;
 pub(crate) use cancellation::enhanced_cancelled_response;
 
 use super::*;
-use request_cancellation::{
-    finalize_cancellation_state, handle_cancel_notification, register_request_cancellation,
-};
 
 impl LspServer {
     /// Handle a JSON-RPC request
     pub fn handle_request(&self, request: JsonRpcRequest) -> Option<JsonRpcResponse> {
-        let id = request.id.clone().and_then(|id| if id.is_null() { None } else { Some(id) });
-        let should_respond = id.is_some();
+        let context = preflight::RequestContext::from_request(&request);
 
-        if handle_cancel_notification(self, &request) {
-            return None; // Notifications don't get responses
+        match preflight::prepare_request(self, &request, &context) {
+            preflight::PreflightOutcome::Continue => {}
+            preflight::PreflightOutcome::NotificationHandled => return None,
+            preflight::PreflightOutcome::Respond(response) => return Some(response),
         }
 
-        if let Some(cancelled) = register_request_cancellation(self, id.as_ref(), &request) {
-            return Some(cancelled);
-        }
-
-        if !self.initialized.load(Ordering::Acquire)
-            && self.initialize_requested.load(Ordering::Acquire)
-            && request.method != "initialize"
-            && request.method != "initialized"
-            && request.method != "shutdown"
-            && request.method != "exit"
-        {
-            self.auto_initialize_for_compat(&request.method);
-        }
-
-        let result = match request.method.as_str() {
-            "initialize" => self.handle_initialize_dispatch(request.params),
-            "initialized" => self.handle_initialized_dispatch(),
-            // Compatibility: some lightweight clients send `initialize` and then
-            // immediately issue requests without an explicit `initialized` notification.
-            // Accept those requests once `initialize` has completed successfully.
-            _ if !self.initialize_requested.load(Ordering::Acquire)
-                && request.method != "shutdown"
-                && request.method != "exit" =>
-            {
-                Err(JsonRpcError {
-                    code: -32002, // ServerNotInitialized per LSP spec
-                    message: "Server not initialized".to_string(),
-                    data: None,
-                })
-            }
-            "shutdown" => self.handle_shutdown_dispatch(),
-            "exit" => self.handle_exit_dispatch(),
-            "textDocument/didOpen" => self.handle_did_open_dispatch(request.params),
-            "textDocument/didChange" => self.handle_did_change_dispatch(request.params),
-            "textDocument/didClose" => self.handle_did_close_dispatch(request.params),
-            "textDocument/didSave" => self.handle_did_save_dispatch(request.params),
-            "textDocument/willSave" => self.handle_will_save_dispatch(request.params),
-            "textDocument/willSaveWaitUntil" => {
-                self.handle_will_save_wait_until_dispatch(request.params)
-            }
-            "notebookDocument/didOpen" => self.handle_notebook_did_open_dispatch(request.params),
-            "notebookDocument/didChange" => {
-                self.handle_notebook_did_change_dispatch(request.params)
-            }
-            "notebookDocument/didSave" => self.handle_notebook_did_save_dispatch(request.params),
-            "notebookDocument/didClose" => self.handle_notebook_did_close_dispatch(request.params),
-            "textDocument/completion" => early_cancel_or!(self, id, "textDocument/completion", {
-                self.handle_completion_cancellable_dispatch(request.params, id.as_ref())
-            }),
-            "completionItem/resolve" => self.handle_completion_resolve_dispatch(request.params),
-            "textDocument/hover" => early_cancel_or!(self, id, "textDocument/hover", {
-                self.handle_hover_cancellable_dispatch(request.params, id.as_ref())
-            }),
-            "textDocument/signatureHelp" => self.handle_signature_help_dispatch(request.params),
-            "textDocument/definition" => early_cancel_or!(self, id, "textDocument/definition", {
-                self.handle_definition_cancellable_dispatch(request.params, id.as_ref())
-            }),
-            "textDocument/declaration" => self.handle_declaration_dispatch(request.params),
-            "textDocument/references" => early_cancel_or!(self, id, "textDocument/references", {
-                self.handle_references_dispatch(request.params)
-            }),
-            "textDocument/documentHighlight" => {
-                self.handle_document_highlight_dispatch(request.params)
-            }
-            "textDocument/prepareTypeHierarchy" => {
-                self.handle_prepare_type_hierarchy_dispatch(request.params)
-            }
-            "typeHierarchy/prepare" => {
-                // Alias for deprecated/alternate method string
-                self.handle_prepare_type_hierarchy_dispatch(request.params)
-            }
-            "typeHierarchy/supertypes" => {
-                self.handle_type_hierarchy_supertypes_dispatch(request.params)
-            }
-            "typeHierarchy/subtypes" => {
-                self.handle_type_hierarchy_subtypes_dispatch(request.params)
-            }
-            "textDocument/diagnostic" => self.handle_document_diagnostic_dispatch(request.params),
-            "workspace/diagnostic" => early_cancel_or!(
-                self,
-                id,
-                "workspace/diagnostic",
-                self.handle_workspace_diagnostic_dispatch(request.params)
-            ),
-            "textDocument/prepareRename" => self.handle_prepare_rename_dispatch(request.params),
-            "workspace/symbol" => early_cancel_or!(self, id, "workspace/symbol", {
-                self.handle_workspace_symbols_dispatch(request.params)
-            }),
-            "workspace/symbol/resolve" => {
-                self.handle_workspace_symbol_resolve_dispatch(request.params)
-            }
-            "textDocument/rename" => self.handle_rename_workspace_dispatch(request.params),
-            "textDocument/codeAction" => self.handle_code_action_dispatch(request.params),
-            "codeAction/resolve" => self.handle_code_action_resolve_dispatch(request.params),
-            "textDocument/semanticTokens/full" => {
-                self.handle_semantic_tokens_dispatch(request.params)
-            }
-            "textDocument/inlayHint" => early_cancel_or!(
-                self,
-                id,
-                "textDocument/inlayHint",
-                self.handle_inlay_hints_dispatch(request.params)
-            ),
-            "inlayHint/resolve" => self.handle_inlay_hint_resolve_dispatch(request.params),
-            "textDocument/documentLink" => self.handle_document_links_dispatch(request.params),
-            "documentLink/resolve" => self.handle_document_link_resolve_dispatch(request.params),
-            "textDocument/selectionRange" => self.handle_selection_range_dispatch(request.params),
-            "textDocument/onTypeFormatting" => {
-                self.handle_on_type_formatting_dispatch(request.params)
-            }
-            "textDocument/codeLens" => self.handle_code_lens_dispatch(request.params),
-            "codeLens/resolve" => self.handle_code_lens_resolve_dispatch(request.params),
-            "textDocument/linkedEditingRange" => {
-                self.handle_linked_editing_range_dispatch(request.params)
-            }
-            "textDocument/inlineCompletion" => {
-                self.handle_inline_completion_dispatch(request.params)
-            }
-            "textDocument/perlInlineCompletionStream" => {
-                self.handle_streaming_inline_completion_dispatch(request.params)
-            }
-            "textDocument/inlineValue" => self.handle_inline_value_dispatch(request.params),
-            "textDocument/moniker" => self.handle_moniker_dispatch(request.params),
-            "textDocument/documentColor" => self.handle_document_color_dispatch(request.params),
-            "textDocument/colorPresentation" => {
-                self.handle_color_presentation_dispatch(request.params)
-            }
-            "textDocument/semanticTokens/range" => {
-                self.handle_semantic_tokens_range_dispatch(request.params)
-            }
-            "workspace/executeCommand" => self.handle_execute_command_dispatch(request.params),
-            "textDocument/typeDefinition" => self.handle_type_definition_dispatch(request.params),
-            "textDocument/implementation" => self.handle_implementation_dispatch(request.params),
-            "textDocument/documentSymbol" => self.handle_document_symbol_dispatch(request.params),
-            "textDocument/foldingRange" => self.handle_folding_range_dispatch(request.params),
-            "textDocument/formatting" => self.handle_formatting_dispatch(request.params),
-            "textDocument/rangeFormatting" => self.handle_range_formatting_dispatch(request.params),
-            "textDocument/rangesFormatting" => {
-                self.handle_ranges_formatting_dispatch(request.params)
-            }
-            "textDocument/prepareCallHierarchy" => {
-                self.handle_prepare_call_hierarchy_dispatch(request.params)
-            }
-            "callHierarchy/incomingCalls" => early_cancel_or!(
-                self,
-                id,
-                "callHierarchy/incomingCalls",
-                self.handle_incoming_calls_dispatch(request.params)
-            ),
-            "callHierarchy/outgoingCalls" => early_cancel_or!(
-                self,
-                id,
-                "callHierarchy/outgoingCalls",
-                self.handle_outgoing_calls_dispatch(request.params)
-            ),
-            "perl/showAst" => self.handle_show_ast_dispatch(request.params),
-            "experimental/testDiscovery" => self.handle_test_discovery_dispatch(request.params),
-            "workspace/configuration" => self.handle_configuration_dispatch(request.params),
-            "workspace/didChangeWatchedFiles" => {
-                self.handle_did_change_watched_files_dispatch(request.params)
-            }
-            "workspace/didChangeWorkspaceFolders" => {
-                self.handle_did_change_workspace_folders_dispatch(request.params)
-            }
-            "workspace/didChangeConfiguration" => {
-                self.handle_did_change_configuration_dispatch(request.params)
-            }
-            "$/perl-lsp/clientResponse" => {
-                self.handle_client_response(request.params);
-                Ok(None)
-            }
-            "window/workDoneProgress/cancel" => {
-                self.handle_progress_cancel_dispatch(request.params)
-            }
-            "workspace/willRenameFiles" => self.handle_will_rename_files_dispatch(request.params),
-            "workspace/didRenameFiles" => self.handle_did_rename_files_dispatch(request.params),
-            "workspace/willDeleteFiles" => self.handle_will_delete_files_dispatch(request.params),
-            "workspace/didDeleteFiles" => self.handle_did_delete_files_dispatch(request.params),
-            "workspace/willCreateFiles" => self.handle_will_create_files_dispatch(request.params),
-            "workspace/didCreateFiles" => self.handle_did_create_files_dispatch(request.params),
-            "workspace/applyEdit" => self.handle_apply_edit_dispatch(request.params),
-            "workspace/textDocumentContent" => {
-                self.handle_text_document_content_dispatch(request.params)
-            }
-            "$/setTrace" => self.handle_set_trace_dispatch(request.params),
-            "$/test/slowOperation" => self.handle_slow_operation_dispatch(&id, request.params),
-            _ => {
-                tracing::debug!(method = %request.method, "Method not implemented");
-                // Enhanced error response with comprehensive context
-                Err(enhanced_error(
-                    METHOD_NOT_FOUND,
-                    &format!("Method '{}' not found or not supported", request.method),
-                    "method_not_found",
-                    Some(&request.method),
-                ))
-            }
-        };
-
-        // Check for enhanced cancellation with provider context before cleanup.
-        // This preserves cancelled responses for requests that are interrupted while
-        // handlers are already running.
-        if let Some(cancelled) = finalize_cancellation_state(id.as_ref()) {
-            return Some(cancelled);
-        }
-
-        match result {
-            Ok(Some(result)) if should_respond => {
-                tracing::trace!(method = %request.method, "Sending successful response");
-                Some(JsonRpcResponse {
-                    jsonrpc: "2.0".to_string(),
-                    id,
-                    result: Some(result),
-                    error: None,
-                })
-            }
-            Ok(Some(_)) => {
-                tracing::trace!(method = %request.method, "Request is a notification (id missing), no response");
-                None
-            }
-            Ok(None) => {
-                tracing::trace!(method = %request.method, "Request is a notification, no response");
-                None // Notification, no response
-            }
-            Err(error) if should_respond => {
-                tracing::debug!(method = %request.method, error = ?error, "Sending error response");
-                Some(JsonRpcResponse {
-                    jsonrpc: "2.0".to_string(),
-                    id,
-                    result: None,
-                    error: Some(error),
-                })
-            }
-            Err(error) => {
-                tracing::debug!(method = %request.method, error = ?error, "Suppressed error response for notification request");
-                None
-            }
-        }
+        let routed = self.route_request(request, context.id.clone(), context.should_respond);
+        response::finalize_response(context.id.as_ref(), routed)
     }
 }
